% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{acc_rate}
\alias{acc_rate}
\title{Calculate acceptance rate}
\usage{
acc_rate(mcmc_results, niter = NULL, nburnin)
}
\arguments{
\item{mcmc_results}{results from inla_is or inla_mcmc}

\item{niter}{number of iterations for the MCMC}

\item{nburnin}{number of iterations for the burn-in}
}
\value{
Acceptance rate.
}
\description{
Calculate acceptance rate
}
