% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthweight}
\alias{birthweight}
\title{Birthweight data}
\format{
## `birthweight`
A data frame with 189 rows and 3 columns:
\describe{
  \item{bwt}{birth weight of the child (in kg, centered)}
  \item{lwt}{mother's weight at last menstrual cycle (in kg, centered)}
  \item{smoke}{mother's smoking status (0 = not smoking, 1 = smoking), suspected to be misreported}
}
}
\source{
The dataset is simulated.
}
\usage{
birthweight
}
\description{
A dataset with 189 observations, with the birth weight of a child as the response, and the smoking status of the mother, as well as the mother's weight at the time of the last menstrual cycle as covariates.
}
\keyword{datasets}
