% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{calculate_pi}
\alias{calculate_pi}
\title{Calculate success probability for binary variable depending on covariate(s)}
\usage{
calculate_pi(alpha, z)
}
\arguments{
\item{alpha}{coefficients for the imputation model}

\item{z}{vector (if only one covariate) or matrix of covariates for imputation model.}
}
\value{
a probability vector
}
\description{
Calculate success probability for binary variable depending on covariate(s)
}
\keyword{internal}
