% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{calculate_summary_statistics}
\alias{calculate_summary_statistics}
\title{Calculate summary statistics}
\usage{
calculate_summary_statistics(summary_df, summary_stat, WW_vec)
}
\arguments{
\item{summary_df}{data frame containing summary.fixed for each iteration (for all fixed effects)}

\item{summary_stat}{the name of the summary statistic of interest}
}
\value{
a data frame with the weighted average across all iterations for each fixed effect of the model
}
\description{
Calculate summary statistics
}
\keyword{internal}
