% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_characteristics}
\alias{data_characteristics}
\title{Calculate some charateristics from simulated data}
\usage{
data_characteristics(data)
}
\arguments{
\item{data}{}
}
\value{
A ggplot2 object giving the sample misclassification matrix.
}
\description{
Calculate some charateristics from simulated data
}
