% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{extract_alphas}
\alias{extract_alphas}
\title{Extract all the alpha samples from model object}
\usage{
extract_alphas(mcmc_results, niter = NULL, nburnin)
}
\arguments{
\item{mcmc_results}{results from inla_is or inla_mcmc}

\item{niter}{number of iterations for the MCMC}

\item{nburnin}{number of iterations for the burn-in}
}
\value{
Data frame of all the alpha samples.
}
\description{
Extract all the alpha samples from model object
}
