% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mcmc.R
\name{FC}
\alias{FC}
\title{FC}
\usage{
FC(X, beta, R, a, varB, n, p)
}
\arguments{
\item{X}{Covariate matrix}

\item{beta}{Coefficient vector}

\item{R}{not sure}

\item{a}{not sure}

\item{varB}{Variance of what?}
}
\value{
C (what is C?)
}
\description{
FC
}
\keyword{internal}
