% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mcmc.R
\name{Fm}
\alias{Fm}
\title{Fm}
\usage{
Fm(X, beta, R, a, Y, varB, n, p)
}
\arguments{
\item{X}{Covariate matrix}

\item{beta}{Coefficient vector}

\item{R}{not sure}

\item{a}{not sure}

\item{Y}{Proposed response value}

\item{varB}{Variance of what?}
}
\value{
m (what is m?)
}
\description{
Fm
}
\keyword{internal}
