% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mcmc.R
\name{FWMat}
\alias{FWMat}
\title{FWMat}
\usage{
FWMat(X, beta, n)
}
\arguments{
\item{X}{Covariate matrix}

\item{beta}{Coefficient vector}
}
\value{
WMat (what is WMat?)
}
\description{
FWMat
}
\keyword{internal}
