% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mcmc.R
\name{FyTilde}
\alias{FyTilde}
\title{FyTilde}
\usage{
FyTilde(X, beta, Y, Ntrials = 1, n)
}
\arguments{
\item{X}{Covariate matrix}

\item{beta}{Coefficient vector}

\item{Y}{Proposed response value}

\item{Ntrials}{Number of trials (for when?)}
}
\value{
yTilde (what is yTilde?)
}
\description{
FyTilde
}
\keyword{internal}
