% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_misclassified.R
\name{generate_misclassified}
\alias{generate_misclassified}
\title{Simulate data with a misclassified covariate}
\usage{
generate_misclassified(
  n,
  p = 2,
  MC_matrix,
  betas = c(1, 2, 2),
  alphas = c(0.5, 0.2),
  family_y = "rnorm",
  sd_y = 1
)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of covariates}

\item{MC_matrix}{misclassification matrix}

\item{betas}{vector of coefficients for the model of interest (the order is 1. intercept, 2. coefficient for variable with error, 3. coefficient for variable without error)}

\item{alphas}{vector of coefficients for the imputation model (the order is 1. intercept, 2. coefficient for variable without error)}

\item{family_y}{likelihood family for the model of interest}

\item{sd_y}{standard deviation for the response variable}
}
\value{
A data frame with a response variable, a continuous covariate, a binary covariate with misclassification, the correct version of that binary covariate, and the probabilities associated with the binary covariate.
}
\description{
Simulate data with a misclassified covariate
}
