% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_vars}
\alias{identify_vars}
\title{Identify roles of covariates in formula}
\usage{
identify_vars(formula_moi, formula_imp)
}
\arguments{
\item{formula_moi}{formula for the model of interest}

\item{formula_imp}{formula for the imputation model}
}
\value{
A list of named variables.
}
\description{
Identify roles of covariates in formula
}
\keyword{internal}
