% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_is.R
\name{inla_is_conditional}
\alias{inla_is_conditional}
\title{Fit model with misclassified covariate with INLA and importance sampling, where misclassification is conditional on response value}
\usage{
inla_is_conditional(
  formula_moi,
  formula_imp = NULL,
  alpha,
  MC_0,
  MC_1,
  data,
  niter,
  ncores = 4,
  ...
)
}
\arguments{
\item{formula_moi}{formula for the model of interest}

\item{formula_imp}{formula for the imputation model}

\item{alpha}{initial values for coefficients for imputation model}

\item{MC_0}{misclassification matrix for y = 0}

\item{MC_1}{misclassification matrix for y = 1}

\item{data}{data for the model}

\item{niter}{number of iterations for the MCMC}

\item{ncores}{number of cores for the parallel computation.}

\item{...}{further arguments to be passed to inla().}
}
\value{
A data frame with the resulting model.
}
\description{
Fit model with misclassified covariate with INLA and importance sampling, where misclassification is conditional on response value
}
