% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{inla_is_misclass}
\alias{inla_is_misclass}
\title{Fit model with misclassified covariate with INLA and importance sampling}
\usage{
inla_is_misclass(
  formula_moi,
  formula_imp = NULL,
  alpha,
  MC_matrix,
  data,
  niter,
  ncores = 4,
  conditional = NULL,
  missing_only = FALSE,
  ...
)
}
\arguments{
\item{formula_moi}{formula for the model of interest}

\item{formula_imp}{formula for the imputation model}

\item{alpha}{values for coefficients of imputation model}

\item{MC_matrix}{misclassification matrix}

\item{data}{data for the model}

\item{niter}{number of iterations for the MCMC}

\item{ncores}{number of cores for the parallel computation.}

\item{conditional}{if the misclassification probabilities are conditional on some variable, then this should be the name of that variable.}

\item{missing_only}{TRUE/FALSE indicating that there is no misclassification, only missingness in a covariate. By default FALSE. Note that if there is misclassification AND missingness, this should still be FALSE.}

\item{...}{further arguments to be passed to inla().}
}
\value{
A data frame with the resulting model.
}
\description{
Fit model with misclassified covariate with INLA and importance sampling
}
