% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mcmc.R
\name{inla_mcmc}
\alias{inla_mcmc}
\title{Fit model with misclassified covariate with INLA and MCMC}
\usage{
inla_mcmc(
  formula_moi,
  formula_imp = NULL,
  alpha,
  MC_matrix,
  data,
  niter = 1600,
  nburnin = 100
)
}
\arguments{
\item{formula_moi}{formula for the model of interest}

\item{formula_imp}{formula for the imputation model}

\item{alpha}{initial values for coefficients for imputation model}

\item{MC_matrix}{misclassification matrix}

\item{data}{data for the model}

\item{niter}{number of iterations for the MCMC}

\item{nburnin}{number of burn-in iterations}
}
\value{
A data frame with the resulting model.
}
\description{
Fit model with misclassified covariate with INLA and MCMC
}
