% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{make_results_df}
\alias{make_results_df}
\title{Summarize results from inla_is_misclass}
\usage{
make_results_df(is_results, niter = length(is_results))
}
\arguments{
\item{is_results}{results from inla_is_misclass}

\item{niter}{number of iterations for the importance sampling}
}
\value{
A dataframe with summary statistics.
}
\description{
Summarize results from inla_is_misclass
}
