% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_misclassified.R
\name{misclassify}
\alias{misclassify}
\title{Generate misclassification in one covariate}
\usage{
misclassify(data, MC_matrix)
}
\arguments{
\item{data}{a data frame from the function `generate_data()`.}

\item{MC_matrix}{misclassification matrix}
}
\value{
A data frame that also contains a column with the misclassified binary covariate.
}
\description{
Generate misclassification in one covariate
}
\keyword{internal}
