% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mcmc.R
\name{new_pi}
\alias{new_pi}
\title{Calculation of new probabilities for x, given the alpha (and rest)}
\usage{
new_pi(alpha, z, MC_matrix, w)
}
\arguments{
\item{alpha}{coefficients for the imputation model}

\item{z}{vector (if only one covariate) or matrix of covariates for imputation model.}

\item{MC_matrix}{misclassification matrix}

\item{w}{vector with misclassified covariate}
}
\value{
a vector with probabilities
}
\description{
Calculation of new probabilities for x, given the alpha (and rest)
}
