% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{new_pi_conditional}
\alias{new_pi_conditional}
\title{Calculation of new probabilities for x, given the alpha (and rest)}
\usage{
new_pi_conditional(alpha, z, MC_matrix, w, conditioning_var)
}
\arguments{
\item{alpha}{coefficients for the imputation model}

\item{z}{vector of covariates for imputation model, and that the misclassification matrix is conditioned on.}

\item{MC_matrix}{list of two misclassification matrices, one for conditioning_var = 0, and the other for conditioning_var = 1.}

\item{w}{vector with misclassified covariate}

\item{conditioning_var}{variable that the misclassification is conditional on}
}
\value{
a vector with probabilities
}
\description{
Calculation of new probabilities for x, given the alpha (and rest)
}
