% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot_compare_inlamisclass}
\alias{plot_compare_inlamisclass}
\title{Plot and compare inlamisclass models with other INLA models}
\usage{
plot_compare_inlamisclass(
  mcmc_results,
  niter = NULL,
  nburnin = 0,
  naive_mod,
  plot_intercept = FALSE,
  num_inlamisclass_models = 1
)
}
\arguments{
\item{mcmc_results}{an inlamisclass model or list of inlamisclass models.}

\item{niter}{number of iterations for the MCMC}

\item{nburnin}{number of iterations for the burn-in}

\item{naive_mod}{a model that does not account for misclassification}

\item{plot_intercept}{should estimated intercept be plotted?}

\item{num_inlamisclass_models}{number of inlamisclass models if mcmc_results is a list}
}
\value{
A ggplot2 object.
}
\description{
Plot and compare inlamisclass models with other INLA models
}
