% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot_inlamisclass}
\alias{plot_inlamisclass}
\title{Plot inlamisclass models}
\usage{
plot_inlamisclass(is_results, niter = length(is_results))
}
\arguments{
\item{is_results}{results from inla_is_misclass}

\item{niter}{number of iterations for the importance sampling}
}
\value{
A ggplot2 object.
}
\description{
Plot inlamisclass models
}
