% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{data_characteristics}
\alias{data_characteristics}
\title{Calculate sample misclassification matrices from simulated data}
\usage{
data_characteristics(data)
}
\arguments{
\item{data}{a simulated data set with a misclassified covariate}
}
\value{
A ggplot2 object giving the sample misclassification matrix.
}
\description{
Calculate sample misclassification matrices from simulated data
}
