% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_is.R
\name{inla_is}
\alias{inla_is}
\title{Fit model with misclassified covariate with INLA and importance sampling}
\usage{
inla_is(data, niter, sample_pi)
}
\arguments{
\item{data}{data for model}

\item{niter}{number of iterations to run}

\item{sample_pi}{probabilities to sample the misclassified covariate from.}
}
\value{
A data frame with the resulting model.
}
\description{
Fit model with misclassified covariate with INLA and importance sampling
}
