% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_pi_conditional}
\alias{new_pi_conditional}
\title{Calculation of new probabilities for x, given the alpha (and rest)}
\usage{
new_pi_conditional(alpha, z, MC_matrix, w)
}
\arguments{
\item{alpha}{coefficients for the imputation model}

\item{z}{vector of covariates for imputation model, and that the misclassification matrix is conditioned on.}

\item{w}{vector with misclassified covariate}

\item{MC_1}{misclassification matrix for z = 1}

\item{MC_0}{misclassification matrix for z = 0}
}
\value{
a vector with probabilities
}
\description{
Calculation of new probabilities for x, given the alpha (and rest)
}
