% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot_compare_inlamisclass}
\alias{plot_compare_inlamisclass}
\title{Plot and compare inlamisclass models with other INLA models}
\usage{
plot_compare_inlamisclass(
  is_results,
  naive_mod,
  correct_mod = NULL,
  niter = NULL,
  plot_intercept = FALSE,
  num_inlamisclass_models = 1
)
}
\arguments{
\item{is_results}{results from inla_is_misclass}

\item{naive_mod}{a model that does not account for misclassification}

\item{correct_mod}{a model that uses the true values of the covariate with misclassification (only possible in simulation studies)}

\item{niter}{number of iterations for the importance sampling}

\item{plot_intercept}{should estimated intercept be plotted?}

\item{num_inlamisclass_models}{number of inlamisclass models if is_results is a list}
}
\value{
A ggplot2 object.
}
\description{
Plot and compare inlamisclass models with other INLA models
}
