% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot_inlamisclass}
\alias{plot_inlamisclass}
\title{Plot inlamisclass models}
\usage{
plot_inlamisclass(
  mcmc_results,
  niter = length(mcmc_results),
  nburnin = length(mcmc_results) - niter,
  naive_mod
)

plot_inlamisclass(
  mcmc_results,
  niter = length(mcmc_results),
  nburnin = length(mcmc_results) - niter,
  naive_mod
)
}
\arguments{
\item{mcmc_results}{results from inla_is or inla_mcmc}

\item{niter}{number of iterations for the MCMC}

\item{nburnin}{number of iterations for the burn-in}

\item{naive_mod}{a model that does not account for misclassification}
}
\value{
A ggplot2 object.

A ggplot2 object.
}
\description{
Plot inlamisclass models

Plot and compare inlamisclass models with other INLA models
}
